+++ tests/testthat/test-apply_base_algorithm.R
@@ -1,7 +1,7 @@
 context("apply_base_algorithm")
 
 test_that("Warning for wrong algorithm choice", {
-  expect_warning(
+  expect_error(
     apply_base_algorithm(tsAirgap, algorithm = "wrongAlgorithm")
   )
 })