+++ tests/testthat/test-na_seasplit.R
@@ -63,7 +63,7 @@ test_that("Imputation works for data.frame", {
 
 
 test_that("Warning for wrong input for algorithm parameter", {
-  expect_warning(na_seasplit(tsAirgap, algorithm = "wrong"))
+  expect_error(na_seasplit(tsAirgap, algorithm = "wrong"))
 })